requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
Django==2.0
django-debug-toolbar==1.11
django-debug-toolbar-template-profiler==1.1.0
django-extensions==2.2.1
graphviz==0.13
gunicorn==19.9.0
idna==2.8
oauthlib==3.1.0
Pillow==6.1.0
pip-tools==4.2.0
psycopg2==2.8.3
pydotplus==2.0.2
PyJWT==1.7.1
pyparsing==2.4.2
python-memcached==1.59
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.3
wrapt==1.11.2
